jellyfish.rb: if we have custom error handler, we should use it
regardless whether handle_exceptions is set or not. otherwise,
we should forbid defining error handler whenever handle_exceptions
is set, which is not easy to do. so i think simply changing the
rule to that if we have custom error handlers, use them first;
otherwise see if we're raising exceptions.