introduce controller_include and change controller definition, see:
so we change the `controller` definition to get a list of
`controller_include`, and create a new class which is a subclass
of `Jellyfish::Controller` and include all modules there.

then we can have this short hand `controller_include` which
remembers what modules we want to include to the automatically
generated controller class.

this greatly reduces client code if we're not doing something
special. (i.e. no helper methods needed)