give a friendly error message if the response is not valid
we only accept:

* Rack triple
* Jellyfish::Response
* nil (then fallback to default in the controller)

we could consider extend this to also accept a body string as in
sinatra, but i don't like the idea of making it even more polymorphic
and unclear about the type. i also consider stop accepting rack triple,
however this might need much more work, and it would be an incompatible
change.

so we would probably make it stop accepting rack triple in the future,
rather than also accepting a body string.