Make sure puts is public
In Muack 1.7.0, visibility will follow with the originally defined
method. In the case of `puts`, it's a private method from Kernel,
which is also available for Array. Here it's supposed to be an IO
object, which `puts` is not the one from Kernel, and it's public.
We need to adopt that so it's turned to public.